feat(ds5): unify controller page in pixel style
